+++ lib/ruby_smb/dcerpc/sec_trailer.rb
@@ -9,9 +9,17 @@ module RubySMB
 
       ndr_uint8  :auth_type
       ndr_uint8  :auth_level
-      ndr_uint8  :auth_pad_length
+      ndr_uint8  :auth_pad_length, initial_value: -> { get_auth_pad_length(@obj) }
       ndr_uint8  :auth_reserved
       ndr_uint32 :auth_context_id
+
+      def get_auth_pad_length(obj)
+        parent = obj&.parent&.parent
+        if parent&.respond_to?(:auth_pad)
+          return parent.auth_pad.length if parent.auth_pad.respond_to?(:length)
+        end
+        0
+      end
     end
 
   end
